reorganize layout and components
