+++ app/_layouts/AccountLayout/index.ts
@@ -1,0 +1,3 @@
+import AccountLayout from './AccountLayout';
+
+export default AccountLayout;
