+++ app/_layouts/Container/index.ts
@@ -1,0 +1,3 @@
+import Container from './Container';
+
+export default Container;
