+++ app/account/orders/[orderId]/page.tsx
@@ -2,7 +2,7 @@ import { redirect } from 'next/navigation';
 
 import OrderDetails from './_components/OrderDetails';
 
-import AccountLayout from '~/account/AccountLayout';
+import AccountLayout from '~/_layouts/AccountLayout';
 
 import { getUserInfo } from '~/lib/SwellGraphQL';
