+++ app/account/orders/page.tsx
@@ -4,7 +4,8 @@ import { BiShoppingBag } from 'react-icons/bi';
 import OrderCard from './_components/OrderCard';
 
 import Button from '~/_components/button/Button';
-import AccountLayout from '~/account/AccountLayout';
+
+import AccountLayout from '~/_layouts/AccountLayout';
 
 import { getUserInfo } from '~/lib/SwellGraphQL';
